fix: Call to "performIdentifierAction()" not working as expected after SDK version bump
Fixed by replacing the call to "performIdentifierAction()" with "setSelectedItemId()" of BottomNavigationView passing in the Id of the MenuItem that needs to be selected. May be there is some code change on how the MenuItem selection is handled. In this updated version, selection process is delegated to a presenter. So, calling to "performIdentifierAction()" did not select/change to the selected MenuItem as it internally calls to the same method "performItemAction()" without passing anything ("null") for the presenter argument.